requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
aiosqlite
python-multipart
openai
python-dotenv
